[change] Updated the serializers imports #287
- Moved the serializer classes of device api
  to a new `serializers.py` file.
- Pointed the install of `openwisp-controller` module
  to the branch containing the new changes.

closes #287
